% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoyant-package.R
\docType{package}
\name{buoyant-package}
\alias{buoyant}
\alias{buoyant-package}
\title{buoyant: Deploy '_server.yml' Compliant Applications to 'DigitalOcean'}
\description{
Provides tools to deploy R web server applications that follow the '_server.yml' standard. This standard allows different R server frameworks ('plumber2', 'fiery', etc.) to be deployed using a common interface. The package supports deployment to 'DigitalOcean' and includes validation tools to ensure '_server.yml' files are correctly formatted.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://posit-dev.github.io/buoyant}
  \item \url{https://github.com/posit-dev/buoyant}
  \item Report bugs at \url{https://github.com/posit-dev/buoyant/issues}
}

}
\author{
\strong{Maintainer}: Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
