% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{create_template}
\alias{create_template}
\title{Create a _server.yml template application}
\usage{
create_template(
  path,
  engine = c("plumber2", "fiery", "generic"),
  overwrite = FALSE
)
}
\arguments{
\item{path}{The directory where the template should be created. Will be created
if it doesn't exist.}

\item{engine}{The engine to use. One of "plumber2", "fiery", or "generic".}

\item{overwrite}{If \code{TRUE}, will overwrite existing files. If \code{FALSE} (default),
will error if files already exist.}
}
\value{
Invisibly returns the path to the created template.
}
\description{
Copies a template _server.yml application to a specified directory.
This is useful for quickly starting a new project with the correct structure.
}
\examples{
\dontrun{
# Create a plumber2 template
create_template("my-api", engine = "plumber2")

# Create a fiery template
create_template("my-fiery-app", engine = "fiery")
}
}
