% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital-ocean.R
\name{do_install_server_deps}
\alias{do_install_server_deps}
\title{Install server dependencies on a droplet}
\usage{
do_install_server_deps(droplet, ...)
}
\arguments{
\item{droplet}{The DigitalOcean droplet that you want to provision
(see \code{\link[analogsea:droplet]{analogsea::droplet()}}). If empty, a new DigitalOcean server will be created.}

\item{...}{Arguments passed into the \code{\link[analogsea:droplet_create]{analogsea::droplet_create()}} function.}
}
\description{
Installs R and common dependencies needed for running _server.yml applications.
This is called automatically by \code{\link[=do_provision]{do_provision()}} but can be called separately
if needed.
}
