% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital-ocean.R
\name{do_ip}
\alias{do_ip}
\title{Get the URL to a deployed application}
\usage{
do_ip(droplet, path)
}
\arguments{
\item{droplet}{The droplet on which to act.}

\item{path}{Optional path to append to the IP address. If not provided, just
returns the IP address.}
}
\value{
A character string with the URL or IP address.
}
\description{
Returns the URL to access a deployed application or the droplet's IP address.
}
