% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital-ocean.R
\name{do_keyfile}
\alias{do_keyfile}
\title{Get the default DigitalOcean SSH keyfile path}
\usage{
do_keyfile()
}
\value{
A character string with the path to the SSH private key, or NULL
if no key is found.
}
\description{
Returns the path to the SSH private key for "digitalocean.com" from
\code{\link[ssh:ssh_credentials]{ssh::ssh_key_info()}}. This is used as the default keyfile for all
buoyant functions that interact with DigitalOcean droplets.
}
\examples{
\dontrun{
  # Get the default keyfile path
  do_keyfile()
}
}
