% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital-ocean.R
\name{do_remove_forward}
\alias{do_remove_forward}
\title{Remove root forwarding}
\usage{
do_remove_forward(droplet, ...)
}
\arguments{
\item{droplet}{The droplet on which to act.}

\item{...}{additional arguments to pass to \code{\link[analogsea:droplet_ssh]{analogsea::droplet_ssh()}}.}
}
\description{
Removes any root path forwarding configuration.
}
