% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{list_templates}
\alias{list_templates}
\title{List available templates}
\usage{
list_templates()
}
\value{
A character vector of available template names.
}
\description{
Lists all available _server.yml templates that can be used with \code{\link[=create_template]{create_template()}}.
}
\examples{
list_templates()
}
